pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "csidrl"
version = "0.1.0"
description = ""
authors = ["Richard Hajek <richard.m.hajek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
matplotlib = "*"
numpy = "*"
pathfinding = "*"
plotly = "*"
jupyterlab = "^3.5.0"
tqdm = "^4.64.1"
qattools = "^0.5.1"
colorama = "^0.4.6"
gym = {version = ">=0.26.0,<0.27.0", extras = ["classic-control"]}
ipython = "^7"
torch = "1.13.1"
# blackbox-backprop = {git = "https://github.com/martius-lab/blackbox-backprop.git"}
# blossom-v = {git = "https://gitlab.tuebingen.mpg.de/mrolinek/blossom_python.git"}
# protobuf = "3.20.*"
# ray = "^2.1.0"
# openpyxl = "^3.0.10"
# pywfc = "^0.0.0"
# diskcache = "^5.4.0"
pytest = "^7.3.0"
tensorboard = "^2.12.2"
diskcache = "^5.6.1"
scikit-learn = "^1.2.2"
stable-baselines3 = ">=1.0.0,<2.0.0"
discord-py = "^2.2.3"
